fix: Check deletion request for an answer that has already been deleted.
이미 삭제한 answer에 대한 요청인지 확인합니다.

+++ src/main/java/com/codessquad/qna/service/AnswerService.java
@@ -34,6 +34,9 @@ public class AnswerService {
         if (!answer.isWriter(tryToDelete)) {
             throw new NotAuthorizedException();
         }
+        if(answer.isDeleted()){
+            throw new NotFoundException(answerId+"번 Answer는 이미 삭제되었습니다.");
+        }
         answer.delete();
     }
